Update to Task API v0.9.0 and some renaming
