+++ requirements.txt
@@ -39,7 +39,7 @@ eventlet==0.24.1
 fs==2.4.4
 Golem-Messages==3.9.0
 Golem-Smart-Contracts-Interface==1.10.2
-golem_task_api==0.7.1
+golem_task_api==0.9.0
 greenlet==0.4.15
 h2==3.0.1
 hpack==3.0.0
