+++ requirements_to-freeze.txt
@@ -18,7 +18,7 @@ eth-utils==1.0.3
 ethereum==1.6.1
 Golem-Messages==3.9.0
 Golem-Smart-Contracts-Interface==1.10.2
-golem_task_api==0.7.1
+golem_task_api==0.9.0
 html2text==2018.1.9
 humanize==0.5.1
 incremental==17.5.0
